requirements.txt
Django==4.1.4
django-filter==21.1
djangorestframework==3.13.1
django-extensions==3.1.5
django-cors-headers==3.11.0
